Large refactor
Moving towards better self-described commands. I made all of the commands static, so hopefully that will work. I didn't see a reason for them to be instances which also helped cleaned up some redundant code before we move to auto-loading.

Moved core libraries into a directory called library. This way all that should be in the commands folder should be commands. Hopefully the next step will be to auto-load the commands in the commands folder.
