+++ library/command.js
@@ -1,0 +1,12 @@
+
+class Command {
+  constructor() {
+    
+  }
+
+  static argsErrorMessage() {
+    return 'Arguments are missing.\nRefer to `!help` or ask an Administrator if this error occurs.';
+  }
+}
+
+module.exports = Command;
